api/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1

# Data validation
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0.post1

# HTTP client
httpx==0.27.0

# Document processing
pypdf==4.1.0
python-docx==1.1.0
python-magic==0.4.27

# AI/ML (lightweight versions for production)
openai==1.13.3
langchain==0.1.11
langchain-openai==0.0.8
langchain-huggingface==0.0.3
sentence-transformers>=2.6.0
faiss-cpu==1.8.0

# Background tasks
celery==5.3.6
redis==5.0.3

# Monitoring & Logging
structlog==24.1.0

# Testing (dev dependencies)
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.0.0
# httpx-test==0.27.0  # Package not found

# Development tools
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
